+++ components/config-provider/configProvider.tsx
@@ -1,41 +1,23 @@
 import {
     defineComponent,
     reactive,
-    PropType,
     provide,
     inject,
     toRef,
     Ref,
     computed,
     ComputedRef,
+    getCurrentInstance,
 } from 'vue';
 import { isObject, isNil } from 'lodash-es';
 
-import type { Theme } from '../_theme/interface';
-import type { GetContainer } from '../_util/interface';
-import {  CONFIG_PROVIDER_INJECTION_KEY } from './const';
 import { hasOwn } from '../_util/utils';
-import type { TypeLanguage } from '../locales';
-import { TypeConfigProviderContext } from './interface';
+import { CONFIG_PROVIDER_INJECTION_KEY, configProviderProps, TypeConfigProviderContext } from './const';
 import { provideLocale } from './useLocale';
-
-export const localeProps = {
-    locale: {
-        type: Object as PropType<TypeLanguage>,
-    },
-};
 
 const globalConfig = reactive<TypeConfigProviderContext>({
     getContainer: () => document.body,
 });
-
-export const configProviderProps = {
-    ...localeProps,
-
-    getContainer: Function as PropType<GetContainer>,
-    theme: String,
-    themeOverrides: Object as PropType<Partial<Theme>>,
-} as const;
 
 // 全局设置
 export const setConfig = (data: TypeConfigProviderContext) => {
@@ -63,7 +45,7 @@ const setProvideConfig = (data: TypeConfigProviderContext) => {
 
 /**
  * 组件中获取配置项，若有局部设置，则取局部设置，若无局部设置，则返回全局设置，为了兼容以下场景：
- * 
+ *
  *  <config-provider :locale="lang1">
  *      locale for here is lang1.
  *      <config-provider :locale="lang2">
@@ -74,16 +56,23 @@ const setProvideConfig = (data: TypeConfigProviderContext) => {
 export function getConfig<K extends keyof TypeConfigProviderContext>(
     key: K,
 ): Ref<TypeConfigProviderContext[K]>;
+
 export function getConfig(): ComputedRef<TypeConfigProviderContext>;
+
 export function getConfig(key?: keyof TypeConfigProviderContext) {
+    // 当不在vue实例使用时
+    const vm = getCurrentInstance();
+    if (!vm) {
+        return computed(() => globalConfig);
+    }
     const providerConfig = inject(CONFIG_PROVIDER_INJECTION_KEY, null);
     const config = computed(() => {
         if (isObject(providerConfig)) {
-            return {...globalConfig, ...providerConfig};
+            return { ...globalConfig, ...providerConfig };
         } else {
             return globalConfig;
         }
-    })
+    });
 
     if (key) {
         if (isObject(providerConfig) && hasOwn(providerConfig, key)) {
@@ -92,7 +81,7 @@ export function getConfig(key?: keyof TypeConfigProviderContext) {
             return toRef(globalConfig, key);
         }
     } else {
-        return config
+        return config;
     }
 }
 
@@ -102,8 +91,8 @@ export default defineComponent({
     setup(props, { slots }) {
         // 兼容子组件局部设置
         setProvideConfig(props);
-        
-        provideLocale(props)
+
+        provideLocale(props, getConfig());
 
         return () => slots.default?.();
     },